Removed the redundant code
